+++ post-backend/parcel-service/src/main/java/com/jowk/parcel/catalog/dto/AdditionalServiceSummary.java
@@ -16,7 +16,7 @@ public record AdditionalServiceSummary(
         return new AdditionalServiceSummary(
                 additionalService.getId(),
                 additionalService.getName(),
-                additionalService.getPrice()
+                additionalService.getPrice().toBigDecimal()
         );
     }
